Split auth server helpers
